EmptyBucket + Seperated ModBucket from ModLiquid file
